misplaced repr_html tests
